meta-mender-tegra/meta-mender-tegra-common/recipes-bsp/u-boot/libubootenv-fake/fw_printenv: fix SC3057 warnings
In dash, or POSIX sh, string indexing isn't supported.
While one solution is to use cut, this isn't possible as cut provided by
coreutils is 1-index based, and the output of nvbootctrl get-current-slot
or tegra-boot-control --current-slot is 0 index based.

The simple solution is to remove the string indexing entirely, and set _cfglbl
to a default value of '"RootfsPartA"', and change it to '"RootfsPartB"' if the
_current_slot is set to 1.

Signed-off-by: Adam Duskett <aduskett@gmail.com>