kas: fix qemuarm64-bootloader-validation
The bootloader validation script requires the
"mender-prepopulate-inactive-partition" feature to be present
for A/B switch testing. As this feature was not available in
upstream meta-mender on the scarthgap branch yet, the revision
was pinned.

This led to build breakage after version bumps, and can be
removed as not necessary anymore.

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>